projects/chess-ml/requirements.txt
python-chess
numpy
pandas
matplotlib
